+++ plugin/src/lib.rs
@@ -25,13 +25,13 @@
 pub use cni::Cni;
 pub use command::Command;
 pub use inputs::Inputs;
-pub use logger::install_logger;
 
 pub mod config;
 #[cfg(any(feature = "with-smol", feature = "with-tokio"))]
 pub mod delegation;
 pub mod error;
 pub mod ip_range;
+pub mod logger;
 pub mod macaddr;
 pub mod reply;
 
@@ -39,6 +39,5 @@ mod cni;
 mod command;
 mod dns;
 mod inputs;
-mod logger;
 mod path;
 mod version;